Add old_readme.html locally (git- and build-ignore it).
This file was emailed by Ervan and shows how to use the package. It's a bit out of date but shold be a good guide to understand what the package should do at this point in time. The file is ignored because it doesn't need to be tracked (.html is read only) and because it's pretty heavy (near 1 mega).